[UMPNPMGR] Fix GCC build
Addendum to 420c4a7abfe14decc624aa9626e8871c9de267e6

+++ base/services/umpnpmgr/install.c
@@ -369,7 +369,7 @@ DeviceInstallThread(LPVOID lpParameter)
 
     while (TRUE)
     {
-        UINT32 devListSize;
+        ULONG devListSize;
         DWORD status = PNP_GetDeviceListSize(NULL, NULL, &devListSize, 0);
         if (status != CR_SUCCESS)
         {
